backend/requirements.txt
fastapi
uvicorn[standard]
pillow
requests
python-multipart
aiofiles
